refactor(commands): extract /ultrareview command into ultrareview.rs
